scripts/python/requirements.txt
boto3
mysql-connector-python
